% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_first.R
\name{check_first}
\alias{check_first}
\title{Perform several checks and fail early.}
\usage{
check_first(
  app_name,
  semantic_version,
  function_name,
  nodejs_path,
  package_install_opts,
  build_path,
  cran_like_url,
  git_host,
  git_repo,
  local_package_path
)
}
\arguments{
\item{app_name}{app_name}

\item{semantic_version}{semantic_version}

\item{function_name}{function_name}

\item{nodejs_path}{nodejs_path}

\item{package_install_opts}{package_install_opts}

\item{build_path}{build_path}

\item{cran_like_url}{cran_like_url}

\item{git_host}{git_host}

\item{git_repo}{git_repo}

\item{local_package_path}{local_package_path}
}
\description{
Perform several checks and fail early.
}
