% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nodejs_npm.R
\name{check_nodejs_npm}
\alias{check_nodejs_npm}
\title{Check if Nodejs and npm works}
\usage{
check_nodejs_npm(os, node_top_dir)
}
\arguments{
\item{os}{os}

\item{node_top_dir}{directory containing nodejs app/exe}
}
\value{
path of nodejs executable if found and functional, otherwise: FALSE
}
\description{
Check if Nodejs and npm works
}
