% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{copy_shinybox_package}
\alias{copy_shinybox_package}
\title{Copy {shinybox} package to an isolated folder.
   This is necessary to avoid dependency-install issues}
\usage{
copy_shinybox_package()
}
\value{
path of new {shinybox}-only library
}
\description{
Copy {shinybox} package to an isolated folder.
   This is necessary to avoid dependency-install issues
}
