% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_r.R
\name{install_r}
\alias{install_r}
\title{Install R into shinybox folder}
\usage{
install_r(
  os,
  cran_like_url = NULL,
  build_path,
  mac_file = NULL,
  mac_r_url = NULL
)
}
\arguments{
\item{os}{os}

\item{cran_like_url}{CRAN-like url e.g. https://cran.r-project.org/bin/windows/base}

\item{build_path}{path to current shinybox app build}

\item{mac_file}{file path to mac OS tar.gz}

\item{mac_r_url}{mac R installer url}
}
\description{
Install R into shinybox folder
}
