% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{install_user_app}
\alias{install_user_app}
\title{Install shiny app package and dependencies}
\usage{
install_user_app(
  library_path = NULL,
  repo_location = "github",
  repo = "chasemc/IDBacApp",
  repos = cran_like_url,
  package_install_opts = NULL,
  rtools_path_win = NULL
)
}
\arguments{
\item{library_path}{path to the new Electron app's R's library folder}

\item{repo_location}{{remotes} package function, one of c("github", "gitlab", "bitbucket", "local")}

\item{repo}{e.g. if repo_location is github: "chasemc/demoApp" ; if repo_location is local: "C:/Users/chase/demoApp"}

\item{repos}{cran like repository package dependencies will be retrieved from}

\item{package_install_opts}{further arguments to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{rtools_path_win}{path to RTools (Windows)}
}
\value{
App name
}
\description{
Install shiny app package and dependencies
}
