% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r_portable_mac.R
\name{make_r_portable_mac}
\alias{make_r_portable_mac}
\title{Change fixed paths to make R portable on macOS}
\usage{
make_r_portable_mac(r_executable_path)
}
\arguments{
\item{r_executable_path}{path to ~bin/R}
}
\value{
NA, modifies file
}
\description{
Change fixed paths to make R portable on macOS
}
