% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybox.R
\name{shinybox}
\alias{shinybox}
\title{shinybox}
\usage{
shinybox(
  app_name = "HAL9000",
  author = "Stanley",
  description = "Heuristically Programmed ALgorithmic Computer",
  semantic_version = "v9000.0.0",
  mran_date = NULL,
  cran_like_url = NULL,
  mac_file = NULL,
 
    mac_r_url = "https://mac.r-project.org/high-sierra/R-4.0-branch/x86_64/R-4.0-branch.tar.gz",
  git_host = "github",
  git_repo = "ocelhay/shinyboxtestapp",
  function_name = "run_app",
  local_package_path = NULL,
  package_install_opts = NULL,
  build_path = NULL,
  rtools_path_win = NULL,
  nodejs_path = file.path(system.file(package = "shinybox"), "nodejs"),
  nodejs_version = "v12.16.2",
  permission = FALSE,
  run_build = TRUE
)
}
\arguments{
\item{app_name}{This will be the name of the executable. It's a uniform type identifier (UTI)
that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters.
see https://www.electron.build/configuration/configuration 
and https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070}

\item{author}{Author of the shiny app}

\item{description}{short app description}

\item{semantic_version}{semantic version of your app, as character (not numeric!);
See https://semver.org/ for more info on semantic versioning.}

\item{mran_date}{MRAN snapshot date, formatted as 'YYYY-MM-DD'}

\item{cran_like_url}{url to cran-like repository}

\item{mac_file}{file path to mac OS tar.gz. Might need to use option 1 of https://www.technipages.com/macos-disable-appname-cant-be-opened-because-it-is-from-an-unidentified-developer}

\item{mac_r_url}{url to mac OS tar.gz}

\item{git_host}{one of c("github", "gitlab", "bitbucket")}

\item{git_repo}{GitHub/Bitbucket/GitLab username/repo of the shiny-app package (e.g. 'ocelhay/shinyboxtestapp').}

\item{function_name}{the function name in your package that starts the shiny app}

\item{local_package_path}{path to local shiny-app package, if 'git_package' isn't used}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{build_path}{Path where the build files will be created, preferably points to an empty directory.
Must not contain a folder with the name as what you put for shinybox(app_name).}

\item{rtools_path_win}{path to RTools (Windows)}

\item{nodejs_path}{path to nodejs}

\item{nodejs_version}{nodejs version to install}

\item{permission}{automatically grant permission to install nodejs and R}

\item{run_build}{logical, whether to start the build process, helpful if you want to modify anthying before building}
}
\description{
shinybox
}
