% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_first.R
\name{check_first}
\alias{check_first}
\title{Perform several checks and fail early}
\usage{
check_first(
  build_path,
  cran_like_url,
  mran_date,
  git_host,
  git_repo,
  local_package_path
)
}
\arguments{
\item{build_path}{build_path}

\item{cran_like_url}{cran_like_url}

\item{mran_date}{mran_date}

\item{git_host}{git_host}

\item{git_repo}{git_repo}

\item{local_package_path}{local_package_path}
}
\description{
Perform several checks and fail early
}
