% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_user_app.R
\name{install_user_app}
\alias{install_user_app}
\title{Install shiny app package and dependencies}
\usage{
install_user_app(
  os,
  library_path,
  repo_location,
  repo,
  repos,
  package_install_opts,
  rtools_path_win
)
}
\arguments{
\item{os}{os}

\item{library_path}{path to the new Electron app's R's library folder}

\item{repo_location}{{remotes} package function, one of c("github", "gitlab", "bitbucket", "local")}

\item{repo}{e.g. if repo_location is github: "ocelhay/shinyboxtestapp" ; if repo_location is local: "C:/Users/olivier/shinyboxtestapp"}

\item{repos}{cran like repository package dependencies will be retrieved from}

\item{package_install_opts}{further arguments to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{rtools_path_win}{path to RTools (Windows)}
}
\value{
App name
}
\description{
Install shiny app package and dependencies
}
