% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_build_release.R
\name{run_build_release}
\alias{run_build_release}
\title{Create an electron-builder release}
\usage{
run_build_release(
  os,
  nodejs_path = file.path(system.file(package = "shinybox"), "nodejs"),
  build_path
)
}
\arguments{
\item{os}{os}

\item{nodejs_path}{parent folder of node.exe (~nodejs_path/node.exe)}

\item{build_path}{path to new electron app top directory}
}
\value{
nothing, used for side-effects
}
\description{
Create an electron-builder release
}
