% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybox.R
\name{shinybox}
\alias{shinybox}
\title{shinybox}
\usage{
shinybox(
  app_name = "HAL",
  author = "Stanley K.",
  description = "Heuristically Programmed ALgorithmic Computer",
  semantic_version = "v9000.0.0",
  cran_like_url = "https://cran.r-project.org",
  mac_file = NULL,
  mac_r_url = NULL,
  git_host = "github",
  git_repo = "ocelhay/shinyboxtestapp",
  local_package_path = NULL,
  function_name = "run_app",
  package_install_opts = NULL,
  build_path = NULL,
  rtools_path_win = NULL,
  nodejs_path = NULL,
  run_build = TRUE
)
}
\arguments{
\item{app_name}{Must be lowercase and one word, and may contain hyphens and underscores.}

\item{author}{Author of the app.}

\item{description}{Short app description.}

\item{semantic_version}{Must be in the form x.x.x and follow the semantic versioning guidelines - https://docs.npmjs.com/about-semantic-versioning.}

\item{cran_like_url}{url to cran-like repository. Default value is https://cran.r-project.org. 
To use MRAN, set to "https://cran.microsoft.com/snapshot/2021-01-10" with desired date. (TODO: what about RStudio package repository?)}

\item{mac_file}{file path to mac OS tar.gz. Might need to use option 1 of https://www.technipages.com/macos-disable-appname-cant-be-opened-because-it-is-from-an-unidentified-developer}

\item{mac_r_url}{url to mac OS tar.gz  (https://mac.r-project.org/)}

\item{git_host}{one of c("github", "gitlab", "bitbucket")}

\item{git_repo}{GitHub/Bitbucket/GitLab username/repo of the shiny-app package (e.g. 'ocelhay/shinyboxtestapp').}

\item{local_package_path}{path to local shiny-app package, if 'git_package' isn't used}

\item{function_name}{the function name in your package that starts the shiny app}

\item{package_install_opts}{optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local}

\item{build_path}{Path where the build files will be created, preferably points to an empty directory.}

\item{rtools_path_win}{path to RTools (Windows only)}

\item{nodejs_path}{path to nodejs}

\item{run_build}{logical, whether to start the build process, helpful if you want to modify anthying before building}
}
\description{
shinybox
}
